fix(vrack-services): fix tests and move order mocks
ref: MANAGER-0000
