+++ package.json
@@ -96,6 +96,7 @@
     "lerna": "^5.5.0",
     "lint-staged": "^13.0.3",
     "lodash": "^4.17.11",
+    "msw": "2.1.7",
     "npm-run-all": "^4.1.5",
     "ora": "^3.4.0",
     "p-series": "^2.1.0",
