+++ packages/manager/apps/vrack-services/e2e/done/create.feature
@@ -14,20 +14,13 @@ Feature: Configuration page
 
   Scenario Outline: User wants to create a new vRack Services
     Given User wants to create a vRack Services with name "<name>" and region <region>
-    Given The order service for vRack Services is <orderVsOk>
-    Given The order service for vRack is <orderVrackOk>
     When User fills the configuration form and clicks the submit button
     Then A modal appear to ask if the user wants to create a new vRack
     When User <acceptOrDenyOrCancel>
     Then User "<returnListing>" on the Listing page
-    Then User sees <anyErrorMessage> error message
 
     Examples:
-      | name | region | orderVsOk | orderVrackOk | acceptOrDenyOrCancel | returnListing  | anyErrorMessage |
-      | vs-1 | RBX    | OK        | OK           | accepts              | returns        | no              |
-      | vs-2 | BHS    | KO        | OK           | accepts              | doesn't return | an              |
-      | vs-3 | LIM    | OK        | KO           | accepts              | returns        | no              |
-      |      | RBX    | KO        | KO           | accepts              | doesn't return | an              |
-      |      | RBX    | OK        | KO           | denies               | returns        | no              |
-      |      | RBX    | OK        | OK           | denies               | returns        | no              |
-      |      | RBX    | OK        | OK           | cancel               | doesn't return | no              |
+      | name | region | acceptOrDenyOrCancel | returnListing  |
+      | vs-1 | RBX    | accepts              | returns        |
+      | vs-2 | BHS    | denies               | returns        |
+      |      | LIM    | cancel               | doesn't return |
