+++ packages/manager/apps/vrack-services/mock/vrack-services/region.ts
@@ -1,4 +1,4 @@
-import { Handler } from '../../e2e/utils/msw-helpers';
+import { Handler } from '@playwright-helpers';
 
 const regionList = [
   {
