+++ packages/manager/apps/vrack-services/package.json
@@ -37,7 +37,6 @@
     "clsx": "^1.2.1",
     "i18next": "^20.4.0",
     "i18next-http-backend": "^2.4.2",
-    "msw": "2.0.11",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-i18next": "^11.18.1",
@@ -45,15 +44,14 @@
     "tailwindcss": "^3.3.3"
   },
   "devDependencies": {
-    "@cucumber/cucumber": "^10.0.1",
+    "@cucumber/cucumber": "^10.3.1",
     "@ovh-ux/manager-vite-config": "*",
-    "@playwright/test": "^1.34.3",
+    "@playwright/test": "^1.41.2",
     "@testing-library/react": "^14.1.2",
     "@testing-library/user-event": "^14.5.1",
-    "@types/cucumber": "^7.0.0",
     "@types/jest": "27.x",
     "jest": "27.x",
-    "jest-cucumber": "^3.0.1",
+    "msw": "2.1.7",
     "ts-jest": "27.x",
     "ts-loader": "^9.5.1",
     "ts-node": "^10.9.1",
