+++ packages/manager/apps/vrack-services/tsconfig.json
@@ -20,6 +20,7 @@
     "baseUrl": ".",
     "paths": {
       "@/*": ["./src/*"],
+      "@playwright-helpers": ["../../../../playwright-helpers/index"],
       "@playwright-helpers/*": ["../../../../playwright-helpers/*"]
     }
   },
